Move error fragment to request
