+++ client/packages/requisitions/src/ResponseRequisition/api/operations.graphql
@@ -1,19 +1,3 @@
-fragment RequisitionReasonNotProvidedError on RequisitionReasonNotProvided {
-  __typename
-  requisitionLine {
-    id
-  }
-  description
-}
-
-fragment RequisitionReasonsNotProvidedError on RequisitionReasonsNotProvided {
-  __typename
-  errors {
-    ...RequisitionReasonNotProvidedError
-  }
-  description
-}
-
 mutation updateResponse(
   $storeId: String!
   $input: UpdateResponseRequisitionInput!
